support java serialization
